split out category_model
